+++ pennylane/workflow/__init__.py
@@ -47,4 +47,4 @@ from .construct_tape import construct_tape
 from .execution import INTERFACE_MAP, SUPPORTED_INTERFACE_NAMES, execute
 from .get_best_diff_method import get_best_diff_method
 from .resolve_diff_method import _resolve_diff_method
-from .qnode import QNode, qnode
+from .qnode import QNode, qnode, _setup_transform_program
