+++ pennylane/workflow/execution.py
@@ -23,18 +23,16 @@ differentiation support.
 
 import inspect
 import logging
-import warnings
-from collections.abc import Callable, MutableMapping
+from collections.abc import Callable
 from functools import partial
 from typing import Literal, Optional, Union, get_args
 from warnings import warn
 
-from cachetools import Cache, LRUCache
+from cachetools import Cache
 
 import pennylane as qml
-from pennylane.tape import QuantumScript, QuantumScriptBatch
-from pennylane.transforms import transform
-from pennylane.typing import Result, ResultBatch
+from pennylane.tape import QuantumScriptBatch
+from pennylane.typing import ResultBatch
 
 from .jacobian_products import DeviceDerivatives, DeviceJacobianProducts, TransformJacobianProducts
 
@@ -93,16 +91,6 @@ INTERFACE_MAP = dict(zip(get_args(SupportedInterfaceUserInput), _mapping_output)
 
 SUPPORTED_INTERFACE_NAMES = list(INTERFACE_MAP)
 """list[str]: allowed interface strings"""
-
-_CACHED_EXECUTION_WITH_FINITE_SHOTS_WARNINGS = (
-    "Cached execution with finite shots detected!\n"
-    "Note that samples as well as all noisy quantities computed via sampling "
-    "will be identical across executions. This situation arises where tapes "
-    "are executed with identical operations, measurements, and parameters.\n"
-    "To avoid this behaviour, provide 'cache=False' to the QNode or execution "
-    "function."
-)
-"""str: warning message to display when cached execution is used with finite shots"""
 
 
 def _use_tensorflow_autograph():
@@ -173,9 +161,7 @@ def _get_ml_boundary_execute(
     return ml_boundary
 
 
-def _make_inner_execute(
-    device, cache, inner_transform, execution_config=None, numpy_only=True
-) -> Callable:
+def _make_inner_execute(device, inner_transform, execution_config=None) -> Callable:
     """Construct the function that will execute the tapes inside the ml framework registration
     for the 1st order derivatives.
 
@@ -197,15 +183,7 @@ def _make_inner_execute(
             cache (None | MutableMapping): The cache to use. If ``None``, caching will not occur.
         """
 
-        transform_program = qml.transforms.core.TransformProgram(inner_transform)
-
-        if numpy_only:
-            transform_program.add_transform(qml.transforms.convert_to_numpy_parameters)
-
-        if cache is not None:
-            transform_program.add_transform(_cache_transform, cache=cache)
-
-        transformed_tapes, transform_post_processing = transform_program(tapes)
+        transformed_tapes, transform_post_processing = inner_transform(tapes)
 
         if transformed_tapes:
             results = device.execute(transformed_tapes, execution_config=execution_config)
@@ -215,43 +193,6 @@ def _make_inner_execute(
         return transform_post_processing(results)
 
     return inner_execute
-
-
-@transform
-def _cache_transform(tape: QuantumScript, cache: MutableMapping):
-    """Caches the result of ``tape`` using the provided ``cache``.
-
-    .. note::
-
-        This function makes use of :attr:`.QuantumTape.hash` to identify unique tapes.
-    """
-
-    def cache_hit_postprocessing(_results: ResultBatch) -> Result:
-        result = cache[tape.hash]
-        if result is not None:
-            if tape.shots and getattr(cache, "_persistent_cache", True):
-                warnings.warn(_CACHED_EXECUTION_WITH_FINITE_SHOTS_WARNINGS, UserWarning)
-            return result
-
-        raise RuntimeError(
-            "Result for tape is missing from the execution cache. "
-            "This is likely the result of a race condition."
-        )
-
-    if tape.hash in cache:
-        return [], cache_hit_postprocessing
-
-    def cache_miss_postprocessing(results: ResultBatch) -> Result:
-        result = results[0]
-        cache[tape.hash] = result
-        return result
-
-    # Adding a ``None`` entry to the cache indicates that a result will eventually be available for
-    # the tape. This assumes that post-processing functions are called in the same order in which
-    # the transforms are invoked. Otherwise, ``cache_hit_postprocessing()`` may be called before the
-    # result of the corresponding tape is placed in the cache by ``cache_miss_postprocessing()``.
-    cache[tape.hash] = None
-    return [tape], cache_miss_postprocessing
 
 
 def _get_interface_name(tapes, interface):
@@ -457,38 +398,13 @@ def execute(
         diff_method, grad_on_execution, interface, device, device_vjp, mcm_config, gradient_kwargs
     )
 
-    is_gradient_transform = isinstance(diff_method, qml.transforms.core.TransformDispatcher)
-    transform_program, inner_transform = _make_transform_programs(
-        device, config, inner_transform, transform_program, is_gradient_transform
-    )
+    # pylint: disable=protected-access
+    if transform_program is None or inner_transform is None:
+        transform_program, inner_transform = qml.workflow._setup_transform_program(
+            transform_program, device, config, cache, cachesize
+        )
 
-    # If caching is desired but an explicit cache is not provided, use an ``LRUCache``.
-    if cache is True:
-        cache = LRUCache(maxsize=cachesize)
-        setattr(cache, "_persistent_cache", False)
-
-    # Ensure that ``cache`` is not a Boolean to simplify downstream code.
-    elif cache is False:
-        cache = None
-
-    # changing this set of conditions causes a bunch of tests to break.
-    no_interface_boundary_required = interface == "numpy" or config.gradient_method in {
-        None,
-        "backprop",
-    }
-    device_supports_interface_data = no_interface_boundary_required and (
-        interface == "numpy"
-        or config.gradient_method == "backprop"
-        or getattr(device, "short_name", "") == "default.mixed"
-    )
-
-    inner_execute = _make_inner_execute(
-        device,
-        cache,
-        inner_transform,
-        config,
-        numpy_only=not device_supports_interface_data,
-    )
+    inner_execute = _make_inner_execute(device, inner_transform, config)
 
     # moved to its own explicit step so that it will be easier to remove
     def inner_execute_with_empty_jac(tapes, **_):
@@ -506,6 +422,10 @@ def execute(
         return post_processing(tapes)
 
     # Exiting early if we do not need to deal with an interface boundary
+    no_interface_boundary_required = interface == "numpy" or config.gradient_method in {
+        None,
+        "backprop",
+    }
     if no_interface_boundary_required:
         results = inner_execute(tapes)
         return post_processing(results)
@@ -615,27 +535,6 @@ def execute(
         )
 
     return post_processing(results)
-
-
-def _make_transform_programs(
-    device, config, inner_transform, transform_program, is_gradient_transform
-):
-    """helper function to make the transform programs."""
-
-    # If diff_method is a gradient transform, device preprocessing should happen in
-    # inner execute (inside the ml boundary).
-    if is_gradient_transform:
-        if inner_transform is None:
-            inner_transform = device.preprocess(config)[0]
-        if transform_program is None:
-            transform_program = qml.transforms.core.TransformProgram()
-    else:
-        if inner_transform is None:
-            inner_transform = qml.transforms.core.TransformProgram()
-        if transform_program is None:
-            transform_program = device.preprocess(config)[0]
-
-    return transform_program, inner_transform
 
 
 def _get_execution_config(
