+++ tests/workflow/test_cache_transform.py
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 """
-Unit tests for the :func:`_cache_transform` and :func:`_apply_cache_transform` functions.
+Unit tests for the :func:`_cache_transform` function.
 """
 # pylint: disable=protected-access,redefined-outer-name
 from collections.abc import MutableMapping
@@ -22,7 +22,7 @@ import pytest
 
 import pennylane as qml
 from pennylane.tape import QuantumScript
-from pennylane.workflow.execution import _cache_transform
+from pennylane.workflow.qnode import _cache_transform
 
 
 @pytest.fixture
@@ -40,7 +40,7 @@ def cache() -> MutableMapping:
 @pytest.fixture
 def transform_spy(mocker) -> MagicMock:
     """Returns a spy on the underlying ``_cache_transform()`` function."""
-    return mocker.spy(qml.workflow.execution._cache_transform, "_transform")
+    return mocker.spy(qml.workflow.qnode._cache_transform, "_transform")
 
 
 def test_cache_miss_before_cache_hit(tape, cache):
